refactor(schedule): extract shouldFinish into a shared module
Move `shouldFinish` + `FinishCheckInput` out of the ExpireEvents job into
`src/lib/schedule/scheduleStatus.ts` so the registration endpoint can reuse
the same "schedule has run out" check without cross-importing a job's
internals. Behaviour is unchanged; the unit cases move alongside it.

Refs #599
